feat(maintner-swpr): refactors and error retries
1. Split code into more logical groups
2. Added Cloud Error Reporting Client to report any errors that occur
3. An error when processing a repository does not abort the entire job,
   rather it logs and reports the error and continues
4. Prints a summary of all errors encountered at the end of the job
