+++ drghs-worker/maintner-swpr/limit_transport.go
@@ -1,0 +1,37 @@
+// Copyright 2019 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package main
+
+import (
+	"net/http"
+
+	"golang.org/x/time/rate"
+)
+
+type limitTransport struct {
+	limiter *rate.Limiter
+	base    http.RoundTripper
+}
+
+func (t limitTransport) RoundTrip(r *http.Request) (*http.Response, error) {
+	limiter := t.limiter
+	if limiter != nil {
+		log.Debug("in limitTransport. Round trip Waiting for limiter")
+		if err := limiter.Wait(r.Context()); err != nil {
+			return nil, err
+		}
+	}
+	return t.base.RoundTrip(r)
+}
